Service testing

+++ src/main/java/com/newbiebe/chatty/repository/MessageRepository.java
@@ -3,5 +3,5 @@ package com.newbiebe.chatty.repository;
 import com.newbiebe.chatty.entity.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepository extends JpaRepository<Message, Integer> {
+public interface MessageRepository extends JpaRepository<Message, Long> {
 }
